Fix Button.stories.js
