+++ client/src/stories/Button.stories.js
@@ -1,36 +1,12 @@
 import React from "react";
 
-import { Button } from "./Button";
+import Button from "../components/Button";
 
 export default {
-  title: "Example/Button",
+  title: "Button",
   component: Button,
-  argTypes: {
-    backgroundColor: { control: "color" },
-  },
 };
 
-const Template = (args) => <Button {...args} />;
+const Template = () => <Button />;
 
 export const Primary = Template.bind({});
-Primary.args = {
-  primary: true,
-  label: "Button",
-};
-
-export const Secondary = Template.bind({});
-Secondary.args = {
-  label: "Button",
-};
-
-export const Large = Template.bind({});
-Large.args = {
-  size: "large",
-  label: "Button",
-};
-
-export const Small = Template.bind({});
-Small.args = {
-  size: "small",
-  label: "Button",
-};
