--- client/src/stories/SubmitButton.stories.js
@@ -1,12 +1,0 @@
-import React from "react";
-
-import SubmitButton from "../components/SubmitButton";
-
-export default {
-  title: "SubmitButton",
-  component: SubmitButton,
-};
-
-const Template = () => <SubmitButton />;
-
-export const Primary = Template.bind({});
